Update command implementation
