+++ platforms/paper/src/main/java/com/badbones69/crazycrates/CrazyCrates.java
@@ -3,9 +3,7 @@ package com.badbones69.crazycrates;
 import com.badbones69.crazycrates.api.FileManager.Files;
 import com.badbones69.crazycrates.api.enums.settings.Messages;
 import com.badbones69.crazycrates.api.managers.quadcrates.SessionManager;
-import com.badbones69.crazycrates.api.objects.CrateLocation;
-import com.badbones69.crazycrates.commands.subs.CrateBaseCommand;
-import com.badbones69.crazycrates.commands.subs.player.BaseKeyCommand;
+import com.badbones69.crazycrates.commands.subs.CommandManager;
 import com.badbones69.crazycrates.cratetypes.CSGO;
 import com.badbones69.crazycrates.cratetypes.Cosmic;
 import com.badbones69.crazycrates.cratetypes.CrateOnTheGo;
@@ -24,31 +22,20 @@ import com.badbones69.crazycrates.listeners.MiscListener;
 import com.badbones69.crazycrates.listeners.PreviewListener;
 import com.badbones69.crazycrates.support.MetricsHandler;
 import com.badbones69.crazycrates.support.libraries.PluginSupport;
-import com.badbones69.crazycrates.support.libraries.UpdateChecker;
 import com.badbones69.crazycrates.support.placeholders.PlaceholderAPISupport;
-import dev.triumphteam.cmd.bukkit.BukkitCommandManager;
-import dev.triumphteam.cmd.bukkit.message.BukkitMessageKey;
-import dev.triumphteam.cmd.core.message.MessageKey;
-import dev.triumphteam.cmd.core.suggestion.SuggestionKey;
-import org.bukkit.command.CommandSender;
 import org.bukkit.configuration.file.FileConfiguration;
-import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.EventPriority;
 import org.bukkit.event.Listener;
 import org.bukkit.event.player.PlayerJoinEvent;
 import org.bukkit.plugin.PluginManager;
 import org.bukkit.plugin.java.JavaPlugin;
-import java.util.ArrayList;
-import java.util.List;
 
 public class CrazyCrates extends JavaPlugin implements Listener {
 
     private static CrazyCrates plugin;
 
     private Starter starter;
-
-    BukkitCommandManager<CommandSender> manager = BukkitCommandManager.create(this);
 
     @Override
     public void onEnable() {
@@ -83,9 +70,6 @@ public class CrazyCrates extends JavaPlugin implements Listener {
 
         boolean metricsEnabled = config.getBoolean("Settings.Toggle-Metrics");
 
-        String updater = config.getString("Settings.Update-Checker");
-        String version = config.getString("Settings.Config-Version");
-
         String menu = config.getString("Settings.Enable-Crate-Menu");
 
         String full = config.getString("Settings.Give-Virtual-Keys-When-Inventory-Full-Message");
@@ -104,12 +88,6 @@ public class CrazyCrates extends JavaPlugin implements Listener {
             Files.CONFIG.saveFile();
         }
 
-        if (version == null) {
-            config.set("Settings.Config-Version", 1);
-
-            Files.CONFIG.saveFile();
-        }
-
         if (menu == null) {
             String oldBoolean = config.getString("Settings.Disable-Crate-Menu");
             boolean switchBoolean = config.getBoolean("Settings.Disable-Crate-Menu");
@@ -124,30 +102,11 @@ public class CrazyCrates extends JavaPlugin implements Listener {
             Files.CONFIG.saveFile();
         }
 
-        if (updater == null) {
-            config.set("Settings.Update-Checker", true);
-
-            Files.CONFIG.saveFile();
-        }
-
-        int configVersion = 1;
-        if (configVersion != config.getInt("Settings.Config-Version") && version != null) {
-            plugin.getLogger().warning("========================================================================");
-            plugin.getLogger().warning("You have an outdated config, Please run the command /crates update!");
-            plugin.getLogger().warning("This will take a backup of your entire folder & update your configs.");
-            plugin.getLogger().warning("Default values will be used in place of missing options!");
-            plugin.getLogger().warning("If you have any issues, Please contact Discord Support.");
-            plugin.getLogger().warning("https://discord.gg/crazycrew");
-            plugin.getLogger().warning("========================================================================");
-        }
-
         if (metricsEnabled) {
             MetricsHandler metricsHandler = new MetricsHandler();
 
             metricsHandler.start();
         }
-
-        checkUpdate();
 
         enable();
     }
@@ -165,33 +124,6 @@ public class CrazyCrates extends JavaPlugin implements Listener {
     public void onPlayerJoin(PlayerJoinEvent e) {
         starter.getCrazyManager().setNewPlayerKeys(e.getPlayer());
         starter.getCrazyManager().loadOfflinePlayersKeys(e.getPlayer());
-    }
-
-    private void checkUpdate() {
-        FileConfiguration config = Files.CONFIG.getFile();
-
-        boolean updaterEnabled = config.getBoolean("Settings.Update-Checker");
-
-        if (!updaterEnabled) return;
-
-        getServer().getScheduler().runTaskAsynchronously(plugin, () -> {
-            UpdateChecker updateChecker = new UpdateChecker(17599);
-
-            try {
-                if (updateChecker.hasUpdate() && !getDescription().getVersion().contains("Beta")) {
-                    getLogger().warning("CrazyCrates has a new update available! New version: " + updateChecker.getNewVersion());
-                    getLogger().warning("Current Version: v" + getDescription().getVersion());
-                    getLogger().warning("Download: " + updateChecker.getResourcePage());
-
-                    return;
-                }
-
-                getLogger().info("Plugin is up to date! - " + updateChecker.getNewVersion());
-            } catch (Exception exception) {
-                getLogger().warning("Could not check for updates! Perhaps the call failed or you are using a snapshot build:");
-                getLogger().warning("You can turn off the update checker in config.yml if on a snapshot build.");
-            }
-        });
     }
 
     public void cleanFiles() {
@@ -237,105 +169,10 @@ public class CrazyCrates extends JavaPlugin implements Listener {
 
         if (PluginSupport.PLACEHOLDERAPI.isPluginEnabled()) new PlaceholderAPISupport().register();
 
-        manager.registerMessage(MessageKey.UNKNOWN_COMMAND, (sender, context) -> sender.sendMessage(Messages.UNKNOWN_COMMAND.getMessage()));
-
-        manager.registerMessage(MessageKey.TOO_MANY_ARGUMENTS, (sender, context) -> {
-            String command = context.getCommand();
-            String subCommand = context.getSubCommand();
-
-            String commandOrder = "/" + command + " " + subCommand + " ";
-
-            String correctUsage = null;
-
-            switch (command) {
-                case "crates" -> correctUsage = getString(subCommand, commandOrder);
-                case "keys" -> {
-                    if (subCommand.equals("view")) correctUsage = "/keys " + subCommand;
-                }
-            }
-
-            if (correctUsage != null) sender.sendMessage(Messages.CORRECT_USAGE.getMessage().replace("%usage%", correctUsage));
-        });
-
-        manager.registerMessage(MessageKey.NOT_ENOUGH_ARGUMENTS, (sender, context) -> {
-            String command = context.getCommand();
-            String subCommand = context.getSubCommand();
-
-            String commandOrder = "/" + command + " " + subCommand + " ";
-
-            String correctUsage = null;
-
-            switch (command) {
-                case "crates" -> correctUsage = getString(subCommand, commandOrder);
-                case "keys" -> {
-                    if (subCommand.equals("view")) correctUsage = "/keys " + subCommand + " <player-name>";
-                }
-            }
-
-            if (correctUsage != null) sender.sendMessage(Messages.CORRECT_USAGE.getMessage().replace("%usage%", correctUsage));
-        });
-
-        manager.registerMessage(MessageKey.INVALID_ARGUMENT, (sender, context) -> sender.sendMessage(Messages.NOT_ONLINE.getMessage().replace("%player%", context.getTypedArgument())));
-
-        manager.registerMessage(BukkitMessageKey.NO_PERMISSION, (sender, context) -> sender.sendMessage(Messages.NO_PERMISSION.getMessage()));
-
-        manager.registerMessage(BukkitMessageKey.PLAYER_ONLY, (sender, context) -> sender.sendMessage(Messages.MUST_BE_A_PLAYER.getMessage()));
-
-        manager.registerMessage(BukkitMessageKey.CONSOLE_ONLY, (sender, context) -> sender.sendMessage(Messages.MUST_BE_A_CONSOLE_SENDER.getMessage()));
-
-        manager.registerSuggestion(SuggestionKey.of("crates"), (sender, context) -> starter.getFileManager().getAllCratesNames(plugin).stream().toList());
-
-        manager.registerSuggestion(SuggestionKey.of("key-types"), (sender, context) -> KEYS);
-
-        manager.registerSuggestion(SuggestionKey.of("online-players"), (sender, context) -> getServer().getOnlinePlayers().stream().map(Player::getName).toList());
-
-        manager.registerSuggestion(SuggestionKey.of("locations"), (sender, context) -> starter.getCrazyManager().getCrateLocations().stream().map(CrateLocation::getID).toList());
-
-        manager.registerSuggestion(SuggestionKey.of("prizes"), (sender, context) -> {
-            List<String> numbers = new ArrayList<>();
-
-            starter.getCrazyManager().getCrateFromName(context.getArgs().get(0)).getPrizes().forEach(prize -> numbers.add(prize.getName()));
-
-            return numbers;
-        });
-
-        manager.registerSuggestion(SuggestionKey.of("numbers"), (sender, context) -> {
-            List<String> numbers = new ArrayList<>();
-
-            for (int i = 1; i <= 250; i++) numbers.add(i + "");
-
-            return numbers;
-        });
-
-        manager.registerArgument(CrateBaseCommand.CustomPlayer.class, (sender, context) -> {
-            return new CrateBaseCommand.CustomPlayer(context);
-        });
-
-        manager.registerCommand(new BaseKeyCommand());
-        manager.registerCommand(new CrateBaseCommand());
+        CommandManager.setup();
 
         printHooks();
     }
-
-    private String getString(String subCommand, String commandOrder) {
-        String correctUsage = null;
-
-        switch (subCommand) {
-            case "transfer" -> correctUsage = commandOrder + "<crate-name> " + "<player-name> " + "<amount>";
-            case "debug", "open", "set" -> correctUsage = commandOrder + "<crate-name>";
-            case "tp" -> correctUsage = commandOrder + "<id>";
-            case "additem" -> correctUsage = commandOrder + "<crate-name> " + "<prize-number>";
-            case "preview", "open-others", "forceopen" -> correctUsage = commandOrder + "<crate-name> " + "<player-name>";
-            case "mass-open" -> correctUsage = commandOrder + "<crate-name> " + "<amount>";
-            case "give-random" -> correctUsage = commandOrder + "<key-type> " + "<amount> " + "<player-name>";
-            case "give", "take" -> correctUsage = commandOrder + "<key-type> " + "<crate-name> " + "<amount> " + "<player-name>";
-            case "giveall" -> correctUsage = commandOrder + "<key-type> " + "<crate-name> " + "<amount>";
-        }
-
-        return correctUsage;
-    }
-
-    private final List<String> KEYS = List.of("virtual", "v", "physical", "p");
 
     public static CrazyCrates getPlugin() {
         return plugin;
