+++ platforms/paper/src/main/java/com/badbones69/crazycrates/commands/subs/admin/CommandSchematics.java
@@ -1,0 +1,23 @@
+package com.badbones69.crazycrates.commands.subs.admin;
+
+import com.badbones69.crazycrates.api.enums.settings.Messages;
+import com.badbones69.crazycrates.commands.subs.CommandManager;
+import dev.triumphteam.cmd.bukkit.annotation.Permission;
+import dev.triumphteam.cmd.core.annotation.SubCommand;
+import org.bukkit.entity.Player;
+import org.bukkit.permissions.PermissionDefault;
+
+public class CommandSchematics extends CommandManager {
+
+    @SubCommand("schem-save")
+    @Permission(value = "crazycrates.command.admin.schematic.save", def = PermissionDefault.OP)
+    public void onAdminSave(Player player) {
+        player.sendMessage(Messages.FEATURE_DISABLED.getMessage());
+    }
+
+    @SubCommand("schem-set")
+    @Permission(value = "crazycrates.command.admin.schematic.set", def = PermissionDefault.OP)
+    public void onAdminSet(Player player) {
+        player.sendMessage(Messages.FEATURE_DISABLED.getMessage());
+    }
+}
